requirements.txt
apache-airflow-providers-amazon
apache-airflow-providers-databricks==2.7.0
apache-airflow-providers-slack==4.2.3
astronomer-providers[databricks]